sample/ update NeteroSurfaceGraphics to use uint32_t as index type

+++ samples/graphicsSurfaces/main.cpp
@@ -29,7 +29,7 @@ std::vector<netero::graphics::Vertex> squareVertices {
     { { -0.5f, 0.5f, 0.f }, { 1.0f, 1.0f, 1.0f }, { 0.f, 1.f } }
 };
 
-std::vector<uint16_t> squareIndices { 0, 1, 2, 2, 3, 0 };
+std::vector<uint32_t> squareIndices { 0, 1, 2, 2, 3, 0 };
 
 std::vector<netero::graphics::Vertex> triangleVertices {
     { { 0.f, -0.2f, 0.f }, { 1.f, 0.0f, 0.f } },
